make the databaseconfig and department name config templates to ease reset

+++ setup/kubernetes-labs/reset-stockmanager-config-department.sh
@@ -1,12 +1,20 @@
 #!/bin/bash
 SCRIPT_NAME=`basename $0`
-if [ $# -eq 0 ]
-  then
-    echo "$SCRIPT_NAME No arguments supplied, you must provide :"
-    echo "  1st the name of your DEPARTMENT, e.g. tg"
-    exit -1 
+if [ -z "$DEFAULT_CLUSTER_CONTEXT_NAME" ]
+then
+  CLUSTER_CONTEXT_NAME=one
+else
+  CLUSTER_CONTEXT_NAME="$DEFAULT_CLUSTER_CONTEXT_NAME"
 fi
-DEPARTMENT=$1
+
+
+if [ $# -ge 0 ]
+then
+  CLUSTER_CONTEXT_NAME=$0
+  echo "$SCRIPT_NAME Operating on supplied context name $CLUSTER_CONTEXT_NAME"
+else
+  echo "$SCRIPT_NAME Using default context name of $CLUSTER_CONTEXT_NAME"
+fi
 
 if [ -z "$AUTO_CONFIRM" ]
 then
@@ -23,25 +31,25 @@ if [ -f $SETTINGS ]
     exit 10
 fi
 
+CONFIG_DIR=$HOME/helidon-kubernetes/configurations/stockmanagerconf/conf
+STOCKMANAGER_CONFIG=$CONFIG_DIR/stockmanager-config.yaml
 if [ "$AUTO_CONFIRM" = "true" ]
 then
   REPLY="y"
-  echo "Auto confirm enabled, Updating the stockmanager config to reset $DEPARTMENT as the department name defaults to $REPLY"
+  echo "Auto confirm enabled, Remove the configured the stockmanager config in $STOCKMANAGER_CONFIG defaults to $REPLY"
 else
-  echo "Updating the stockmanager config to reset $DEPARTMENT as the department name."
+  echo "Remove the configured the stockmanager config in $STOCKMANAGER_CONFIG ."
   read -p "Proceed (y/n) ?" REPLY
 fi
 if [[ ! $REPLY =~ ^[Yy]$ ]]
 then
   echo "OK, exiting"
   exit 1
-else
-    echo "Department reset using $DEPARTMENT as the department name"
 fi
-STOCKMANAGER_CONFIG=$HOME/helidon-kubernetes/configurations/stockmanagerconf/conf/stockmanager-config.yaml
-TEMP="$STOCKMANAGER_CONFIG".tmp
-echo "Updating the stockmanager config in $STOCKMANAGER_CONFIG to reset $DEPARTMENT as the department name"
-# echo command is "s/  department: \"$DEPARTMENT Shop\"/#  department: \"My Shop\"/"
-cat $STOCKMANAGER_CONFIG | sed -e "s/  department: \"$DEPARTMENT Shop\"/#  department: \"My Shop\"/" > $TEMP
-rm $STOCKMANAGER_CONFIG
-mv $TEMP $STOCKMANAGER_CONFIG
+if [ -f "$STOCKMANAGER_CONFIG" ]
+then
+    echo "Removing the configured the stockmanager config in $STOCKMANAGER_CONFIG"
+    rm $STOCKMANAGER_CONFIG
+else
+    echo "Stock mamager config in $STOCKMANAGER_CONFIG already removed"
+fi
